feat(src): Updates to history 3.x.x (#26)
* feat(src): update all dependencies

Update to history 3.x and all dev dependencies for testing

* feat(): updateto history 3

* chore(travis): use node 6

* chore(travis): use chrome

* chore(travis): fix travis config

* chore(travis): use chrome

* chore(travis): use chrome

* chore(travis): trusty ubuntu

* chore(travis): use firefox

* chore(travis): sleep

* chore(travis): before_script

* chore(travis): some fixes maybe

* chore(testem): downgrade testem

* chore(package): only test node

Currently testem ci is broken and is unable to run firefox. Shoud investigate fixes
later, but to get a new release out, that *should* work, I'm taking this shortcut.
