+++ history/.travis.yml
@@ -2,8 +2,7 @@ language: node_js
 addons:
   firefox: "latest"
 node_js:
-  - "4"
-  - "5"
+  - "6"
 before_script:
   - "export DISPLAY=:99.0"
   - "sh -e /etc/init.d/xvfb start"
